rhdl_bits

Module alias

Type Aliases

This crate provides a set of type aliases that you can use to save keystrokes. These type aliases are named b1 through b128 for [Bits], and s1 through s128 for [SignedBits]. So, for example, if you want to represent a 32 bit value, you can use the type alias b32 instead of the full type name [Bits]<32>. For example:

use rhdl_bits::alias::*;
let bits: b32 = 0xDEAD_BEEF.into();
let signed_bits: s4 = (-2).into();

Note that in order to avoid differences in behavior between Rust arithmetic and hardware arithmetic, that you should use b8, b16, b32, b64, b128 instead of u8, u16, u32, u64, u128. Similarly, use s8, s16, s32, s64, s128 instead of i8, i16, i32, i64, i128. Note also that b1 is not the same as bool. b1 is a 1 bit unsigned integer type, and bool is a 1 bit boolean type. The two types are not interchangeable. When making comparisons or using other Rust constructs that require a boolean type, you cannot simply substitute b1. On the other hand, you can use bool in your design instead of b1. The two types behave the same way, even though they are different.

Functions

b1

b2

b3

b4

b5

b6

b7

b8

b9

b10

b11

b12

b13

b14

b15

b16

b17

b18

b19

b20

b21

b22

b23

b24

b25

b26

b27

b28

b29

b30

b31

b32

b33

b34

b35

b36

b37

b38

b39

b40

b41

b42

b43

b44

b45

b46

b47

b48

b49

b50

b51

b52

b53

b54

b55

b56

b57

b58

b59

b60

b61

b62

b63

b64

b65

b66

b67

b68

b69

b70

b71

b72

b73

b74

b75

b76

b77

b78

b79

b80

b81

b82

b83

b84

b85

b86

b87

b88

b89

b90

b91

b92

b93

b94

b95

b96

b97

b98

b99

b100

b101

b102

b103

b104

b105

b106

b107

b108

b109

b110

b111

b112

b113

b114

b115

b116

b117

b118

b119

b120

b121

b122

b123

b124

b125

b126

b127

b128

s2

s3

s4

s5

s6

s7

s8

s9

s10

s11

s12

s13

s14

s15

s16

s17

s18

s19

s20

s21

s22

s23

s24

s25

s26

s27

s28

s29

s30

s31

s32

s33

s34

s35

s36

s37

s38

s39

s40

s41

s42

s43

s44

s45

s46

s47

s48

s49

s50

s51

s52

s53

s54

s55

s56

s57

s58

s59

s60

s61

s62

s63

s64

s65

s66

s67

s68

s69

s70

s71

s72

s73

s74

s75

s76

s77

s78

s79

s80

s81

s82

s83

s84

s85

s86

s87

s88

s89

s90

s91

s92

s93

s94

s95

s96

s97

s98

s99

s100

s101

s102

s103

s104

s105

s106

s107

s108

s109

s110

s111

s112

s113

s114

s115

s116

s117

s118

s119

s120

s121

s122

s123

s124

s125

s126

s127

s128

Type Aliases

b1

b2

b3

b4

b5

b6

b7

b8

b9

b10

b11

b12

b13

b14

b15

b16

b17

b18

b19

b20

b21

b22

b23

b24

b25

b26

b27

b28

b29

b30

b31

b32

b33

b34

b35

b36

b37

b38

b39

b40

b41

b42

b43

b44

b45

b46

b47

b48

b49

b50

b51

b52

b53

b54

b55

b56

b57

b58

b59

b60

b61

b62

b63

b64

b65

b66

b67

b68

b69

b70

b71

b72

b73

b74

b75

b76

b77

b78

b79

b80

b81

b82

b83

b84

b85

b86

b87

b88

b89

b90

b91

b92

b93

b94

b95

b96

b97

b98

b99

b100

b101

b102

b103

b104

b105

b106

b107

b108

b109

b110

b111

b112

b113

b114

b115

b116

b117

b118

b119

b120

b121

b122

b123

b124

b125

b126

b127

b128

s2

s3

s4

s5

s6

s7

s8

s9

s10

s11

s12

s13

s14

s15

s16

s17

s18

s19

s20

s21

s22

s23

s24

s25

s26

s27

s28

s29

s30

s31

s32

s33

s34

s35

s36

s37

s38

s39

s40

s41

s42

s43

s44

s45

s46

s47

s48

s49

s50

s51

s52

s53

s54

s55

s56

s57

s58

s59

s60

s61

s62

s63

s64

s65

s66

s67

s68

s69

s70

s71

s72

s73

s74

s75

s76

s77

s78

s79

s80

s81

s82

s83

s84

s85

s86

s87

s88

s89

s90

s91

s92

s93

s94

s95

s96

s97

s98

s99

s100

s101

s102

s103

s104

s105

s106

s107

s108

s109

s110

s111

s112

s113

s114

s115

s116

s117

s118

s119

s120

s121

s122

s123

s124

s125

s126

s127

s128