rhdl_bits

Struct Bits

pub struct Bits<Len> {
    pub val: u128,
    /* private fields */
}

The Bits type is a fixed-sized bit vector. It is meant to imitate the behavior of bit vectors in hardware. Due to the design of the Bits type, you can only create a Bits type of up to 128 bits in length for now. However, you can easily express larger constructs in hardware using arrays, tuples and structs. The only real limitation of the Bits type being 128 bits is that you cannot perform arbitrary arithmetic on longer bit values in your hardware designs. I don’t think this is a significant issue, but the Bits design of the rust-hdl crate was much slower and harder to maintain and use. I think this is a good trade-off.

Note that the Bits type implements 2’s complement arithmetic. See https://en.wikipedia.org/wiki/Two%27s_complement for more information.

Note also that the Bits kind is treated as an unsigned value for the purposes of comparisons. If you need signed comparisons, you will need the SignedBits type.

Fields
val: u128
Implementations

impl<N: BitWidth> Bits<N>

pub const MASK: Self

Defines a constant Bits value with all bits set to 1.

pub const MAX: Self

pub const ZERO: Self

pub const fn mask() -> Self

Return a Bits value with all bits set to 1.

let bits = Bits::<U8>::mask();
assert_eq!(bits, 0xFF);

pub const fn xext<M>(self) -> Bits<Sum<N, M>>
where N: Add<M>, Sum<N, M>: BitWidth,

Extend a value by the number of bits specified.

pub const fn xshr<M>(self) -> Bits<Diff<N, M>>
where N: Sub<M>, M: BitWidth, Diff<N, M>: BitWidth,

pub const fn xshl<M>(self) -> Bits<Sum<N, M>>
where N: Add<M>, M: BitWidth, Sum<N, M>: BitWidth,

pub const fn resize<M: BitWidth>(self) -> Bits<M>

pub const fn as_signed(self) -> SignedBits<N>

Reinterpret the Bits value as a SignedBits value.

pub const fn raw(self) -> u128

Extract the raw u128 behind the Bits value.

pub const fn dyn_bits(self) -> DynBits

Convert the compile-time sized Bits to a run-time tracked [DynBits] value.

pub fn to_bools(self) -> Vec<bool>

Build a (dynamic, stack allocated) vector containing the bits that make up this value. This will be slow.

pub fn any(self) -> bool

pub fn all(self) -> bool

pub fn xor(self) -> bool